requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
starlette
itsdangerous
jinja2
psycopg2
python-dotenv
